Refactor options initialization logic and fix run type.
